create agent dialogs
