+++ apps/chat/src/components/Common/AgentDialogs.tsx
@@ -1,0 +1,129 @@
+import { useCallback } from 'react';
+
+import { useTranslation } from 'next-i18next';
+
+import { translate } from '@/src/utils/app/translation';
+
+import { DialAIEntityModel } from '@/src/types/models';
+import { SharingType } from '@/src/types/share';
+import { Translation } from '@/src/types/translation';
+
+import { ApplicationActions, ModelsActions } from '@/src/store/actions';
+import { ApplicationSelectors } from '@/src/store/application/application.selectors';
+import { useAppDispatch, useAppSelector } from '@/src/store/hooks';
+import {
+  MarketplaceActions,
+  MarketplaceSelectors,
+} from '@/src/store/marketplace/marketplace.reducers';
+import {
+  PublicationActions,
+  PublicationSelectors,
+} from '@/src/store/publication/publication.reducers';
+
+import { DeleteType } from '@/src/constants/marketplace';
+
+import { PublishModal } from '../Chat/Publish/PublishWizard';
+import { ApplicationLogs } from '../Marketplace/ApplicationLogs';
+import { ConfirmDialog } from './ConfirmDialog';
+
+const getDeleteConfirmationText = (
+  action: DeleteType,
+  entity: DialAIEntityModel,
+) => {
+  const translationVariables = {
+    modelName: entity.name,
+    modelVersion: entity.version
+      ? translate(' (version {{version}})', { version: entity.version })
+      : '',
+  };
+
+  const deleteConfirmationText = {
+    [DeleteType.DELETE]: {
+      heading: translate('Confirm deleting application'),
+      description: translate(
+        'Are you sure you want to delete the {{modelName}}{{modelVersion}}?',
+        translationVariables,
+      ),
+      confirmLabel: translate('Delete'),
+    },
+    [DeleteType.REMOVE]: {
+      heading: translate('Confirm removing agent'),
+      description: translate(
+        'Are you sure you want to remove {{modelName}} from My workspace?',
+        translationVariables,
+      ),
+      confirmLabel: translate('Remove'),
+    },
+  };
+
+  return deleteConfirmationText[action];
+};
+
+export const AgentDialogs = () => {
+  const { t } = useTranslation(Translation.Marketplace);
+
+  const dispatch = useAppDispatch();
+
+  const deleteModel = useAppSelector(MarketplaceSelectors.selectDeleteModel);
+  const publishModel = useAppSelector(PublicationSelectors.selectPublishModel);
+  const logsEntityId = useAppSelector(ApplicationSelectors.selectLogsEntityId);
+
+  const handleDeleteClose = useCallback(
+    (confirm: boolean) => {
+      if (confirm && deleteModel) {
+        if (deleteModel.action === DeleteType.REMOVE) {
+          dispatch(
+            ModelsActions.removeInstalledModels({
+              references: [deleteModel.entity.reference],
+              action: DeleteType.REMOVE,
+            }),
+          );
+        } else if (deleteModel.action === DeleteType.DELETE) {
+          dispatch(ApplicationActions.delete(deleteModel.entity));
+        }
+
+        dispatch(MarketplaceActions.setDetailsModel());
+      }
+
+      dispatch(MarketplaceActions.setDeleteModel());
+    },
+    [deleteModel, dispatch],
+  );
+  const handlePublishClose = useCallback(() => {
+    dispatch(PublicationActions.setPublishModel());
+  }, [dispatch]);
+
+  const handleCloseApplicationLogs = useCallback(() => {
+    dispatch(ApplicationActions.setLogsEntityId());
+  }, [dispatch]);
+
+  return (
+    <>
+      {/* MODALS */}
+      {!!deleteModel && (
+        <ConfirmDialog
+          isOpen
+          {...getDeleteConfirmationText(deleteModel.action, deleteModel.entity)}
+          onClose={handleDeleteClose}
+          cancelLabel={t('Cancel')}
+        />
+      )}
+      {!!(publishModel && publishModel?.entity?.id) && (
+        <PublishModal
+          entity={publishModel.entity}
+          type={SharingType.Application}
+          isOpen={!!publishModel}
+          onClose={handlePublishClose}
+          publishAction={publishModel.action}
+        />
+      )}
+      {logsEntityId && (
+        <ApplicationLogs
+          isOpen
+          onClose={handleCloseApplicationLogs}
+          entityId={logsEntityId}
+        />
+      )}
+    </>
+  );
+};
